requirements.txt
# Core dependencies
streamlit==1.31.0
google-generativeai==0.3.2
python-dotenv==1.0.0

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# OCR and image processing (optional - for scanned PDFs)
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0

# Vector database
faiss-cpu==1.9.0
numpy==1.26.4

# Text processing
tiktoken>=0.7.0

# Utilities
langchain==0.1.0
langchain-community==0.0.13
